#322 create fiscal json test file
